backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI chat backend application"
authors = [
    {name = "Pavlo Leskovych"}
]
license = "Apache-2.0"
requires-python = ">=3.12"
dynamic = ["dependencies"]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
sqlalchemy = "^2.0.45"
alembic = "^1.17.2"
pydantic = "^2.12.5"
python-dotenv = "^1.2.1"
python-multipart = "^0.0.21"
boto3 = "^1.42.21"
websockets = "^15.0.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
# Pin bcrypt < 4.1: passlib expects bcrypt.__about__.__version__, removed in bcrypt 4.1+
bcrypt = ">=4.0.1,<4.1"
pydantic-settings = "^2.12.0"
psycopg2-binary = "^2.9.11"
httpx = "^0.28.1"
sqladmin = "^0.23.0"
itsdangerous = ">=1.1.0"

[tool.poetry.group.dev.dependencies]
black = "^25.12.0"
pre-commit = "^4.5.1"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.pyright]
include = ["app"]
exclude = ["**/__pycache__", "**/.*"]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
